Add files via upload
In today's competitive business landscape, companies are constantly seeking ways to optimize their operations and improve their decision-making processes. One crucial aspect of business optimization is predicting future demand, especially in the retail sector. In this model, we delve into the process of predicting the number of orders a company may receive for a particular product using historical sales data.